% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{category_df}
\alias{category_df}
\title{Category vector}
\format{
\code{category_df}
A data frame with XXXX rows and XXX columns:
\describe{
\item{feature}{all features available for spatial prioritisation}
\item{category}{the categories of all features}
...
}
}
\source{
"Made up data"
}
\usage{
category_df
}
\description{
A tibble containing specific grouping of features
}
\keyword{datasets}
