% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Category_vec2}
\alias{Category_vec2}
\title{Category vector}
\format{
\code{Category_vec2}
A data frame with XXXX rows and XXX columns:
\describe{
\item{feature}{all features available for spatial prioritisation}
\item{category}{the categories of all features}
...
}
}
\source{
"Made up data"
}
\usage{
Category_vec2
}
\description{
A tibble containing specific grouping of features
}
\keyword{datasets}
