% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CoralSeaVelocity}
\alias{CoralSeaVelocity}
\title{Example climate velocity for the example region of the coral sea}
\format{
\code{CoralSeaVelocity}
A data frame with 397 rows and 3 columns:
\describe{
\item{vocMagg_transformed}{Climate velocity column}
\item{geometry}{Geometry column of sf object}
...
}
}
\source{
https://www.iucnredlist.org/
}
\usage{
CoralSeaVelocity
}
\description{
tsf file containing climate velocity (SSP5-8.5) of the coral sea
}
\keyword{datasets}
