% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dat_bathy}
\alias{dat_bathy}
\title{Bathymetry Data}
\format{
\code{dat_bathy}
A data frame with bathymetry data
\describe{
\item{cellID}{ID number for each row (or cell) of the dataset}
\item{bathymetry}{Bathymetry data}
\item{geometry}{sf geometry column}
...
}
}
\source{
"oceandatr package"
}
\usage{
dat_bathy
}
\description{
Bathymetry data downloaded via the oceandatr package
}
\keyword{datasets}
