% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dat_category}
\alias{dat_category}
\title{Category vector}
\format{
\code{dat_category}
A data frame with XXXX rows and XXX columns:
\describe{
\item{feature}{all features available for spatial prioritisation}
\item{category}{the categories of all features}
...
}
}
\source{
"Made up data"
}
\usage{
dat_category
}
\description{
A tibble containing specific grouping of features
}
\keyword{datasets}
