% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dat_clim}
\alias{dat_clim}
\title{Climate metric data}
\format{
\code{dat_clim}
A data frame with 780 rows and 3 columns:
\describe{
\item{geometry}{sf geometry column}
\item{metric}{climate metric information}
...
}
}
\source{
"Made up data"
}
\usage{
dat_clim
}
\description{
An sf dataframe of planning units for spatial prioritisation
}
\keyword{datasets}
