% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dat_mpas}
\alias{dat_mpas}
\title{MPA Data}
\format{
\subsection{\code{dat_mpas}}{

A data frame with XXXX rows and XXX columns:
\describe{
\item{cellID}{ID number for each row (or cell) of the dataset}
\item{geometry}{sf geometry column}
\item{wdpa}{binary MPA information (1: MPA)}
...
}
}
}
\source{
"Made up data"
}
\usage{
dat_mpas
}
\description{
An sf dataframe of planning units for spatial prioritisation
}
\keyword{datasets}
