% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dat_problem}
\alias{dat_problem}
\title{prioritizr Problem}
\format{
\subsection{\code{dat_problem}}{

A class that stores the data, mathematical formula and other information needed to run the prioritisation:
\describe{
\item{data}{list containing information on the cost and features}
\item{objective}{object specifying objective function used in conservation problem}
...
}
}
}
\source{
"Made up data"
}
\usage{
dat_problem
}
\description{
An conservation problem for spatial prioritisation
}
\keyword{datasets}
