% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dat_region}
\alias{dat_region}
\title{Regionalisation}
\format{
\code{dat_region}
A data frame with XXXX rows and XXX columns:
\describe{
\item{geometry}{sf geometry column}
...
}
}
\source{
"Made up data"
}
\usage{
dat_region
}
\description{
An sf dataframe of planning units for spatial prioritisation
}
\keyword{datasets}
