% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dat_soln2}
\alias{dat_soln2}
\title{prioritizr Solution2}
\format{
\code{dat_soln2}
A data frame with XXXX rows and XXX columns:
\describe{
\item{cellID}{ID number for each row (or cell) of the dataset}
\item{geometry}{sf geometry column}
...
}
}
\source{
"Made up data"
}
\usage{
dat_soln2
}
\description{
An sf dataframe of planning units for spatial prioritisation
}
\keyword{datasets}
