% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dat_species_bin2}
\alias{dat_species_bin2}
\title{Binary Species Data 2}
\format{
\code{dat_species_bin2}
A data frame with XXXX rows and XXX columns:
\describe{
\item{geometry}{sf geometry column}
...
}
}
\source{
"Made up data"
}
\usage{
dat_species_bin2
}
\description{
An sf dataframe of planning units for spatial prioritisation
}
\keyword{datasets}
