% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dat_species_prob}
\alias{dat_species_prob}
\title{Probability Species Data}
\format{
\subsection{\code{dat_species_prob}}{

A data frame with XXXX rows and XXX columns:
\describe{
\item{cellID}{ID number for each row (or cell) of the dataset}
\item{geometry}{sf geometry column}
...
}
}
}
\source{
"Made up data"
}
\usage{
dat_species_prob
}
\description{
An sf dataframe of planning units for spatial prioritisation
}
\keyword{datasets}
