% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{IUCN_IA_Targets}
\alias{IUCN_IA_Targets}
\title{IUCN and Inverse Area based Targets for Aquamaps example species}
\format{
\code{IUCN_IA_Targets}
A data frame with 15 rows and 4 columns:
\describe{
\item{Species}{Species names as found in the IUCN database}
\item{Area km2}{total suitable habitat of a species}
\item{Target}{Individual species-specific target}
\item{IUCN_Category}{IUCN category of the species}
...
}
}
\source{
https://www.iucnredlist.org/
}
\usage{
IUCN_IA_Targets
}
\description{
tbl_df file containing IUCN and IA targets of the selected species of the coral sea
}
\keyword{datasets}
