% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MPAsCoralSea}
\alias{MPAsCoralSea}
\title{Current MPAs of the Coral Sea}
\format{
\code{MPAsCoralSea}
A data frame with 397 rows and 17 columns:
\describe{
\item{wdpa}{binary vector denoting presence or absence of a current marine protected area}
...
}
}
\source{
https://www.protectedplanet.net/en/thematic-areas/wdpa?tab=WDPA
}
\usage{
MPAsCoralSea
}
\description{
sf file containing the current marine protected areas of the coral sea
}
\keyword{datasets}
