% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialplanr-package.R
\docType{package}
\name{spatialplanr-package}
\alias{spatialplanr}
\alias{spatialplanr-package}
\title{spatialplanr: Help files for prioritzr Spatial Planning projects}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

This package provides a range of tools for setting up, running and plotting a spatial plan using the prioritizr. It is only intended for use by the Mathematical Marine Ecology Lab at UQ, as it sets up projects using their approach and data sets.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/MathMarEcol/spatialplanr}
  \item Report bugs at \url{https://github.com/MathMarEcol/spatialplanr/issues}
}

}
\author{
\strong{Maintainer}: Jason D. Everett \email{JaseEverett@gmail.com} (\href{https://orcid.org/0000-0002-6681-8054}{ORCID})

Authors:
\itemize{
  \item Sandra Neubert \email{sandra.neubert97@gmail.com} (\href{https://orcid.org/0000-0002-3112-4116}{ORCID})
}

}
\keyword{internal}
