% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatPlan_Extras.R
\name{SpatPlan_ArrangeFeatures}
\alias{SpatPlan_ArrangeFeatures}
\title{Ensure all features are in the same order.}
\usage{
SpatPlan_ArrangeFeatures(df)
}
\arguments{
\item{df}{sf object to sort by Lon and Lat}
}
\value{

}
\description{
Ensure all features are in the same order.
}
