% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatPlan_Convert_2PUs.R
\name{SpatPlan_Convert_2PUs}
\alias{SpatPlan_Convert_2PUs}
\title{Function to interpolate data onto Planning Units}
\usage{
SpatPlan_Convert_2PUs(dat, PlanUnits)
}
\arguments{
\item{dat}{Dataset or filename of dataset}

\item{PlanUnits}{\code{sf} object of Planning Units}
}
\value{
\code{sf} object containing the Planning Units and the feature.
}
\description{
The dataset needs to be raster or vector format.
If the input contains continuous data, the output is an area-averaged mean for each planning unit.
If the input is binary, the output is the proportion of the planning unit covered.
}
