% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatPlan_Convert_PacificRobinson.R
\name{SpatPlan_Convert_PacificRobinson}
\alias{SpatPlan_Convert_PacificRobinson}
\title{Convert a region to Pacific-Centred Robinson Projection.}
\usage{
SpatPlan_Convert_PacificRobinson(df, buff = 0)
}
\arguments{
\item{df}{An 'sf' object in longlat space}

\item{buff}{The buffer to use for}
}
\value{
An \code{sf} object coverted to Pacific-Robinson projection
}
\description{
Convert a region to Pacific-Centred Robinson Projection.
}
