% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatPlan_CrossingPUs.R
\name{SpatPlan_Convert_Poly2PUs}
\alias{SpatPlan_Convert_Poly2PUs}
\title{Cross raw features (as polygons) with Planning Units}
\usage{
SpatPlan_Convert_Poly2PUs(file, PUs, binary = FALSE)
}
\arguments{
\item{file}{A filename to load or a dataframe to work with.}

\item{PUs}{The planning units to be crossed with}

\item{binary}{Is the feature binary (TRUE) or continuous (FALSE)}
}
\value{

}
\description{
At the moment this requires a sf object with only the geometry and 1 data column.
}
\details{
It also assumes the feature has to overlap with the centroid of the PU. This means
some small features won't be included, but also stops us over representing the area of these small features.
}
