% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatPlan_CrossingPUs.R
\name{SpatPlan_Convert_Rast2PUs}
\alias{SpatPlan_Convert_Rast2PUs}
\title{Cross raw features (as prast) with Planning Units}
\usage{
SpatPlan_Convert_Rast2PUs(file, PUs, binary = FALSE)
}
\arguments{
\item{binary}{}
}
\value{

}
\description{
Cross raw features (as prast) with Planning Units
}
\examples{

}
