% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{SpatPlan_Create_Polygon}
\alias{SpatPlan_Create_Polygon}
\title{Function for creating polygon}
\usage{
SpatPlan_Create_Polygon(x, cCRS)
}
\arguments{
\item{x}{A named vector of lon/lat coordinates from which to make an \code{sf} polygon}

\item{cCRS}{The CRS to use for the polygon}
}
\value{
An \code{sf} object for the polygon
}
\description{
Function for creating polygon
}
