% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatPlan_Extras.R
\name{SpatPlan_Create_SinglePolygon}
\alias{SpatPlan_Create_SinglePolygon}
\title{Create ONE BIG polygon that we can use to populate with PUs}
\usage{
SpatPlan_Create_SinglePolygon(df, res)
}
\arguments{
\item{res}{Resolution of the raster in degrees}
}
\value{

}
\description{
Create ONE BIG polygon that we can use to populate with PUs
}
