% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{SpatPlan_Crop_AQM}
\alias{SpatPlan_Crop_AQM}
\title{Crop AquaMaps Data}
\usage{
SpatPlan_Crop_AQM(df, spp, extent)
}
\arguments{
\item{df}{The AquaMaps \code{stars} object from \code{SpatPlan_Get_AquaMaps}}

\item{spp}{A character vector of species}

\item{extent}{An \code{sf} object from which to extract the extent}
}
\value{
An \code{sf} object cropped to \code{extent}
}
\description{
Crop AquaMaps Data
}
