% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatPlan_Extras.R
\name{SpatPlan_fix_FBtype}
\alias{SpatPlan_fix_FBtype}
\title{A function to fix incorrect column types from fishbase and sealifebase}
\usage{
SpatPlan_fix_FBtype(df, datab = "fishbase")
}
\arguments{
\item{df}{The species dataframe}

\item{datab}{Which server to use "fishbase" (default) or "sealifebase"}
}
\value{

}
\description{
A function to fix incorrect column types from fishbase and sealifebase.
This may be able to be removed. These errors should be fixed in a future FB version.
}
