% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatPlan_Get_MPAs.R
\name{SpatPlan_Get_MPAs}
\alias{SpatPlan_Get_MPAs}
\title{Get locked in areas}
\usage{
SpatPlan_Get_MPAs(
  PUs,
  cCRS = "+proj=longlat +datum=WGS84 +ellps=WGS84 +towgs84=0,0,0",
  Status = "Designated",
  Desig = c("National", "Regional", "International", "Not Applicable"),
  Direc = file.path("~", "SpatPlan_Data")
)
}
\arguments{
\item{Desig}{}
}
\value{

}
\description{
Get locked in areas
}
