% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatPlan_Get_PlanningUnits.R
\name{SpatPlan_Get_PlanningUnits}
\alias{SpatPlan_Get_PlanningUnits}
\title{Function to create square or heaxagonal planning units for your area of interest.}
\usage{
SpatPlan_Get_PlanningUnits(
  Bndry,
  InnerB,
  CellArea,
  Shape = "hexagon",
  inverse = FALSE
)
}
\arguments{
\item{Bndry}{An sf polygon object which outlines the limits of the study area.}

\item{InnerB}{An sf multipolygon object which contains all the areas (ie land) that you wish to remove from the grid.}

\item{CellArea}{The area in km you wish your resultant Planning Units to be.}

\item{Shape}{Hexagon or Square}

\item{inverse}{To inver the \code{InnerB} and keep the areas within the InnerB}
}
\value{

}
\description{
The code takes the bbox so the limits are the most important.
The output inherits the crs from this sf object so ensure it is in the correct projection for your needs
The code assumes that any Planning Unit whose centroids is over land will be removed. This approximates > 50\% of the PU is landward.
}
\details{
Written by Jason Everett (UQ/UNSW/CSIRO)
Written: 15 December 2020
Updated: 25th February 2022
}
