% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatPlan_Get_TargetsIA.R
\name{SpatPlan_Get_TargetsIA}
\alias{SpatPlan_Get_TargetsIA}
\title{Calculate Inverse Area Targets}
\usage{
SpatPlan_Get_TargetsIA(df, target_min, target_max)
}
\arguments{
\item{target_max}{}
}
\value{

}
\description{
Calculate Inverse Area Targets
}
