% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{SpatPlan_Match_Names}
\alias{SpatPlan_Match_Names}
\title{Substitute numbers for all_names in regionalisations}
\usage{
SpatPlan_Match_Names(dat, nam)
}
\arguments{
\item{dat}{\code{sf} data frame with one column of numeric/integer corresponding to \code{nam}}

\item{nam}{character vector of names corresponding to numeric column of dat}
}
\value{
An \code{sf} dataframe with numeric regionalisations substituted for category names
}
\description{
Substitute numbers for all_names in regionalisations
}
