% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatPlan_Plotting.R
\name{SpatPlan_Plot_Comparison}
\alias{SpatPlan_Plot_Comparison}
\title{Plot solution comparison}
\usage{
SpatPlan_Plot_Comparison(soln1, soln2, world)
}
\arguments{
\item{world}{}
}
\value{

}
\description{
Plot solution comparison
}
