% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatPlan_Plotting.R
\name{SpatPlan_Plot_Cost}
\alias{SpatPlan_Plot_Cost}
\title{Plot cost}
\usage{
SpatPlan_Plot_Cost(Cost, landmass)
}
\arguments{
\item{Cost}{An \code{sf} object of cost for \code{prioritizr}}

\item{landmass}{An \code{sf} object of land polygon}
}
\value{
A ggplot object of the plot
}
\description{
Plot cost
}
