% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatPlan_Plotting.R
\name{SpatPlan_Plot_FeatureNo}
\alias{SpatPlan_Plot_FeatureNo}
\title{Plot number of features}
\usage{
SpatPlan_Plot_FeatureNo(df, world)
}
\arguments{
\item{world}{}
}
\value{

}
\description{
Plot number of features
}
