% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatPlan_Plotting.R
\name{SpatPlan_Plot_Longhurst}
\alias{SpatPlan_Plot_Longhurst}
\title{Plot Longhurst Provinces}
\usage{
SpatPlan_Plot_Longhurst(PlanUnits, world)
}
\arguments{
\item{world}{}
}
\value{

}
\description{
Plot Longhurst Provinces
}
