% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatPlan_Plotting.R
\name{SpatPlan_Plot_MPAs}
\alias{SpatPlan_Plot_MPAs}
\title{Plot MPAs}
\usage{
SpatPlan_Plot_MPAs(df, landmass)
}
\arguments{
\item{df}{An \code{sf} object of marine protected areas}

\item{landmass}{An \code{sf} object of land polygon}
}
\value{
A ggplot object of the plot
}
\description{
Plot MPAs
}
