% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatPlan_Plotting.R
\name{SpatPlan_Plot_PUs}
\alias{SpatPlan_Plot_PUs}
\title{Plot Planning Units}
\usage{
SpatPlan_Plot_PUs(PlanUnits, landmass)
}
\arguments{
\item{PlanUnits}{Planning Units as an \code{sf} object}

\item{landmass}{An \code{sf} object of land polygon}
}
\value{
A ggplot object of the plot
}
\description{
Plot Planning Units
}
