% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatPlan_Plotting.R
\name{SpatPlan_Plot_Solution}
\alias{SpatPlan_Plot_Solution}
\title{Plot prioritizr solution}
\usage{
SpatPlan_Plot_Solution(s1, PlanUnits, world)
}
\arguments{
\item{world}{}
}
\value{

}
\description{
Plot prioritizr solution
}
