% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatPlan_Process_Geomorphic.R
\name{SpatPlan_Process_Geomorphic}
\alias{SpatPlan_Process_Geomorphic}
\title{Reprocess the geomorphic data to rds to make it easier to use.}
\usage{
SpatPlan_Process_Geomorphic(Direc = file.path("~", "SpatPlan_Data"))
}
\arguments{
\item{Direc}{The directory where the MME data is being stored. If not specified, the default location is assumed.}
}
\description{
Reprocess the geomorphic data to rds to make it easier to use.
}
\examples{

}
