% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{SpatPlan_Replace_NAs}
\alias{SpatPlan_Replace_NAs}
\title{Remove NAs from spatial data using nearest neighbour}
\usage{
SpatPlan_Replace_NAs(df, vari)
}
\arguments{
\item{df}{An \code{sf} dataframe}

\item{vari}{Variable to}
}
\value{
An \code{sf} object with NAs replaced with the nearest neighbour
}
\description{
Remove NAs from spatial data using nearest neighbour
}
