% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{SpatPlan_Scale_01}
\alias{SpatPlan_Scale_01}
\title{Scale spatial layers to between 0 and 1}
\usage{
SpatPlan_Scale_01(dat, col_name)
}
\arguments{
\item{dat}{\code{sf} dataframe}

\item{col_name}{Name of the column to scale}
}
\value{
\code{sf} dataframe
}
\description{
Scale spatial layers to between 0 and 1
}
