% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatPlan_Extras.R
\name{SpatPlan_validate_FBNAs}
\alias{SpatPlan_validate_FBNAs}
\title{Code to loop through and try and validate names with fishbase and sealifebase}
\usage{
SpatPlan_validate_FBNAs(spp, datab = "fishbase")
}
\arguments{
\item{spp}{A dataframe with the species to be checked.}

\item{datab}{Which server to use "fishbase" (default) or "sealifebase"}
}
\value{

}
\description{
Code to loop through and try and validate names with fishbase and sealifebase
}
\examples{
SpatPlan_validate_FBNAs(data.frame(Species = "Thunnus maccoyii"))
}
