% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{spDataFiltered}
\alias{spDataFiltered}
\title{Example....}
\format{
\subsection{\code{spDataFiltered}}{

A data frame with .....
\describe{
}
}
}
\source{
"Made up data"
}
\usage{
spDataFiltered
}
\description{
Example....
}
\keyword{datasets}
