% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-climate.R
\name{splnr_ClimatePriorityArea_CSapproach}
\alias{splnr_ClimatePriorityArea_CSapproach}
\title{Function to split the feature data into climate-smart (CS) and non-climate-smart (NCS) areas depending on the percentile chosen by the user.}
\usage{
splnr_ClimatePriorityArea_CSapproach(
  featuresDF,
  percentile,
  metricDF,
  direction
)
}
\arguments{
\item{featuresDF}{feature sf object which should have a column for cellID}

\item{percentile}{cut-off threshold for determining whether an area is a climate priority area or not (e.g., lower 35th percentile of warming or upper 65th percentile of acidification). Note that the percentile here is the lower limit of the threshold.}

\item{metricDF}{climate metric data.frame with 'metric' as the column name of the metric values per planning unit. This should also have a column for the cellID}

\item{direction}{If direction = 1, metric values are from low (least climate-smart) to high (most climate-smart). If direction = -1, metric values are from high (least climate-smart) to low (most climate-smart).}
}
\value{
A new sf dataframe that has cutoffs applied.
}
\description{
Function to split the feature data into climate-smart (CS) and non-climate-smart (NCS) areas depending on the percentile chosen by the user.
}
\examples{
metric_df <- dat_clim

dat_species_binDF <- dat_species_bin \%>\%
  sf::st_drop_geometry()

out_sf <- splnr_ClimatePriorityArea_CSapproach(
  featuresDF = dat_species_bin,
  percentile = 5, metricDF = metric_df, direction = 1
)
}
