% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{splnr_convert2PacificRobinson}
\alias{splnr_convert2PacificRobinson}
\title{Convert a world Robinson sf object to a Pacific-centred one}
\usage{
splnr_convert2PacificRobinson(df, buff = 0)
}
\arguments{
\item{df}{An sf dataframe}

\item{buff}{The buffer too apply to features that cross after merge}
}
\value{
An sf object in the Robinson projection
}
\description{
Written by Jason D. Everett
UQ/CSIRO/UNSW
Last edited 8th Sept 2021
}
