% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splnr_deprecated.R
\name{splnr_convert_regionalisation}
\alias{splnr_convert_regionalisation}
\title{Function to interpolate regionalisation data onto Planning Units}
\usage{
splnr_convert_regionalisation(dat, PUs, cat_name = NA, col_name = NA)
}
\arguments{
\item{dat}{Dataset in raster or sf format.}

\item{PUs}{\code{sf} object of Planning Units}

\item{cat_name}{A character string of all categories in the regionalisation}

\item{col_name}{The name of the layer}
}
\value{
\code{sf} object containing the Planning Units and the feature.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
}
\details{
This is a wrapper for \code{splnr_Convert2PUs()} but deals with need to processes each layer seperately

The dataset needs to be raster or vector format.
If the input contains continuous data, the output is an area-averaged mean for each planning unit.
If the input is binary, the output is the proportion of the planning unit covered.
}
\examples{
\dontrun{
df <- splnr_convert_regionalisation(dat, PUs)
}
}
