% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{splnr_convert_toPacific}
\alias{splnr_convert_toPacific}
\title{Convert a world sf object to a Pacific-centred one
Defaults to assuming Robinson projection}
\usage{
splnr_convert_toPacific(
  df,
  buff = 0,
  cCRS =
    "+proj=robin +lon_0=180 +x_0=0 +y_0=0 +ellps=WGS84 +datum=WGS84 +units=m +no_defs"
)
}
\arguments{
\item{df}{An sf dataframe}

\item{buff}{The buffer too apply to features that cross after merge}

\item{cCRS}{The crs to use for the output.}
}
\value{
An sf object in the Robinson projection
}
\description{
Written by Jason D. Everett
UQ/CSIRO/UNSW
Last edited 8th Sept 2021
}
\examples{
df_rob <- rnaturalearth::ne_coastline(returnclass = "sf") \%>\%
  splnr_convert_toPacific()
}
