% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-climate.R
\name{splnr_CPA_CSapproach_assignTargets}
\alias{splnr_CPA_CSapproach_assignTargets}
\title{Function to assign targets for climate-priority-area approach}
\usage{
splnr_CPA_CSapproach_assignTargets(
  featuresDF,
  targetsDF,
  climateSmartDF,
  refugiaTarget = 1
)
}
\arguments{
\item{featuresDF}{feature \code{sf}object which should have a column for cellID}

\item{targetsDF}{\code{data.frame}with list of features under "feature" column and their corresponding targets under "target" column}

\item{climateSmartDF}{\code{sf} object produced using the function splnr_ClimatePriorityArea_CSapproach()}

\item{refugiaTarget}{target assigned to climate-smart areas}
}
\value{
A new sf dataframe that has cutoffs applied.
}
\description{
Function to assign targets for climate-priority-area approach
}
\examples{
Features <- dat_species_bin \%>\%
  dplyr::select(-"cellID")

target <- Features \%>\%
  sf::st_drop_geometry() \%>\%
  colnames() \%>\%
  data.frame() \%>\%
  setNames(c("feature")) \%>\%
  dplyr::mutate(target = 0.3)

metric_df <- dat_clim

dat_species_binDF <- dat_species_bin \%>\%
sf::st_drop_geometry()

out_sf <- splnr_ClimatePriorityArea_CSapproach(featuresDF = dat_species_bin,
        percentile = 5, metricDF = metric_df, direction = 1)

target <- splnr_CPA_CSapproach_assignTargets(featuresDF = Features,
                             targetsDF = target,
                             climateSmartDF = out_sf,
                            refugiaTarget = 1)
}
