% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-climate.R
\name{splnr_CS_feature_preprocess}
\alias{splnr_CS_feature_preprocess}
\title{Function to run the feature climate-smart approach}
\usage{
splnr_CS_feature_preprocess(featuresDF, percentile, metricDF, direction)
}
\arguments{
\item{featuresDF}{feature \code{sf}object which should have a column for cellID}

\item{percentile}{cut-off threshold for determining whether an area is a climate priority area or not. Note that the percentile here is the lower limit of the threshold.}

\item{metricDF}{climate metric data.frame with 'metric' as the column name of the metric values per planning unit. This should also have a column for the cellID}

\item{direction}{If direction = 1, metric values are from low (least climate-smart) to high (most climate-smart). If direction = -1, metric values are from high (least climate-smart) to low (most climate-smart).}
}
\value{
A \code{list} with two components: 1. is the data frame passed to \code{prioritizr} when creating a conservation problem containing the binary information per planning unit per feature. 2. are the targets for the features in the conservation problem when the CPA approach is used.
}
\description{
This function creates a climate layer by selecting the most climate-smart areas in the entire planning region.
}
\examples{

metric_df <- dat_clim

featureTest <- splnr_CS_feature_preprocess(featuresDF = dat_species_bin,
  percentile = 5, metricDF = metric_df, direction = 1)
}
