% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{splnr_featureNames}
\alias{splnr_featureNames}
\title{Returns the feature names}
\usage{
splnr_featureNames(dat, exclude = NA)
}
\arguments{
\item{dat}{sf dataframe of features}

\item{exclude}{Character vector of any columes to exclude}
}
\value{
A character vector of names
}
\description{
Returns the feature names
}
\examples{
df <- dat_species_prob \%>\%
 splnr_featureNames(exclude = c("cellID"))
}
