% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splnr_get_nearestDist2Coast.R
\name{splnr_get_nearestDist2Coast}
\alias{splnr_get_nearestDist2Coast}
\title{Function to compute distances to nearest coastline for each centroid of each planning unit in the grid provided.}
\usage{
splnr_get_nearestDist2Coast(grid)
}
\arguments{
\item{grid}{An sf polygon object which contains outlines the limits of the study area.}
}
\value{
An \code{sf} object with distances to the nearest coast
}
\description{
The code takes a sf object and return it updated with a new coastDistance column.
The output inherits the crs from this sf object so ensure it is in the correct projection for your needs
}
\details{
Written by Kristine Buenafe
Written: March/April 2023
Modified by Kilian Barreiro
Updated: December 2023
}
