% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splnr_get_planningUnits.R
\name{splnr_get_PlanningUnits}
\alias{splnr_get_PlanningUnits}
\title{Function to create square or heaxagonal planning units for your area of interest.}
\usage{
splnr_get_PlanningUnits(
  Bndry,
  InnerB,
  CellArea = 1000,
  Shape = "hexagon",
  inverse = FALSE
)
}
\arguments{
\item{Bndry}{An sf polygon object which outlines the limits of the study area.}

\item{InnerB}{An sf multipolygon object which contains all the areas (ie land) that you wish to remove from the grid.}

\item{CellArea}{The area in km you wish your resultant Planning Units to be.}

\item{Shape}{Hexagon or Square}

\item{inverse}{To invert the \code{InnerB} and keep the areas within the InnerB}
}
\value{
An \code{sf} object of planning units
}
\description{
The code takes the bbox so the limits are the most important.
The output inherits the crs from this sf object so ensure it is in the correct projection for your needs
The code assumes that any Planning Unit whose centroids is over land will be removed. This approximates > 50\% of the PU is landward.
}
\details{
Some example options for hexagonal \code{PU_size}
PU_size <- 5240 # km2 (~1 deg at equator)
PU_size <- 2620 # km2 (~0.5 deg at equator)
PU_size <- 669.9 # km2 (~0.25 deg at equator)

Written by Jason Everett (UQ/UNSW/CSIRO)
Written: 15 December 2020
Updated: 8th April 2022
}
\examples{
\dontrun{
dat_bndry_moll <- dat_bndry \%>\% sf::st_transform("ESRI:54009")
landmass <- rnaturalearth::ne_countries(returnclass = "sf") \%>\% sf::st_transform("ESRI:54009")
PUs <- splnr_get_PlanningUnits(dat_bndry_moll, landmass, CellArea = 10000, Shape = "hexagon")
ggplot2::ggplot() +
  ggplot2::geom_sf(data = dat_bndry_moll, colour = "blue", fill = NA) +
  ggplot2::geom_sf(data = PUs, colour = "black", fill = NA)
  }
}
