% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splnr_get_targetsIA.R
\name{splnr_get_TargetsIA}
\alias{splnr_get_TargetsIA}
\title{Calculate Inverse Area Targets}
\usage{
splnr_get_TargetsIA(df, target_min, target_max)
}
\arguments{
\item{df}{An \code{sf} dataframe with features to calculate}

\item{target_min}{The minimum target for inverse area}

\item{target_max}{The maximum target for inverse area}
}
\value{
An \code{sf} dataframe with Inverse Area Targets added in \code{Target}
}
\description{
This function takes a min (\code{target_min}) and max (\code{target_max}) target range and calculates an inverse area target for each feature based on areal coverage.
}
\examples{
targets <- dat_species_prob \%>\%
  splnr_get_TargetsIA(target_min = 0.3, target_max = 0.8)
}
