% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{splnr_match_names}
\alias{splnr_match_names}
\title{Substitute Numbers for Names in Regionalizations}
\usage{
splnr_match_names(dat, nam)
}
\arguments{
\item{dat}{An \code{sf} data frame with a single non-geometry column containing
numeric or integer values that correspond to the names in \code{nam}.}

\item{nam}{A named character vector. The \emph{names} of this vector should be
the numeric/integer values found in \code{dat}'s column, and the \emph{values} of
this vector should be the desired character names for substitution.}
}
\value{
An \code{sf} dataframe where the numeric/integer values in the relevant
column have been substituted with the corresponding character names from \code{nam}.
}
\description{
\code{splnr_match_names()} replaces numeric or integer values in a spatial
(sf) dataframe's column with corresponding character names, typically used
for regionalization data.
}
\details{
This function is designed for scenarios where spatial data contains numeric
identifiers for regions, and you have a mapping (a named character vector)
to convert these IDs into more descriptive names. It assumes that the \code{sf}
dataframe (\code{dat}) has only one non-geometry column that needs recoding.

The function directly applies the mapping from the \code{nam} vector to the
specified column. The names of the \code{nam} vector should correspond to the
numeric/integer values in the \code{dat} column, and the values of \code{nam} will
be the new character names.
}
\examples{
# Define the named character vector for mapping.
region_names <- c("Region1" = "SE Aust", "Region2" = "Tas", "Region3" = "NE Aust")

# Apply the function to substitute numeric codes with names.
df_named_regions <- splnr_match_names(dat = dat_region, nam = region_names)
print(df_named_regions)
}
