% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splnr_plotting.R
\name{splnr_plot_binFeature}
\alias{splnr_plot_binFeature}
\title{Plot binary feature}
\usage{
splnr_plot_binFeature(
  df,
  colInterest,
  colorVals = c(Suitable = "#3182bd", `Not Suitable` = "#c6dbef"),
  showLegend = TRUE,
  plotTitle = " ",
  legendTitle = "Habitat"
)
}
\arguments{
\item{df}{A \verb{data frame} with binary feature information}

\item{colInterest}{column of data frame that contains binary information of feature to plot}

\item{colorVals}{A \code{list} object of named vectors that will match the color value with the according name. "TRUE" stands for selected planning units.}

\item{showLegend}{A logical command on whether to show the legend of the solution (Default: TRUE).}

\item{plotTitle}{A character value for the title of the plot. Can be empty ("").}

\item{legendTitle}{A character value for the title of the legend. Can be empty ("").}
}
\value{
A ggplot object of the plot
}
\description{
Plot binary feature
}
\examples{
splnr_plot_binFeature(dat_species_bin, dat_species_bin$Spp1)
}
