% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splnr_plotting_climate.R
\name{splnr_plot_climKernelDensity_Basic}
\alias{splnr_plot_climKernelDensity_Basic}
\title{Basic Kernel Density Plots for climate-smart spatial plans}
\usage{
splnr_plot_climKernelDensity_Basic(soln)
}
\arguments{
\item{soln}{The \code{prioirtizr} solution containing a "metric" column containing the used climate metric information}
}
\value{
A ggplot object of the plot
}
\description{
Basic Kernel Density Plots for climate-smart spatial plans
}
\examples{
Features <- dat_species_bin \%>\%
  dplyr::select(-"cellID")

target <- Features \%>\%
  sf::st_drop_geometry() \%>\%
  colnames() \%>\%
  data.frame() \%>\%
  setNames(c("feature")) \%>\%
  dplyr::mutate(target = 0.3)

dat_species_binDF <- dat_species_bin \%>\%
  sf::st_drop_geometry()

out_sf <- splnr_ClimatePriorityArea_CSapproach(
  featuresDF = dat_species_bin,
  percentile = 5, metricDF = dat_clim, direction = -1
)

target <- splnr_CPA_CSapproach_assignTargets(
  featuresDF = Features,
  targetsDF = target,
  climateSmartDF = out_sf,
  refugiaTarget = 1
)

out_sf <- out_sf \%>\%
  dplyr::mutate(Cost_None = rep(1, 780)) \%>\%
  dplyr::left_join(dat_clim \%>\%
    sf::st_drop_geometry(), by = "cellID")

usedFeatures <- out_sf \%>\%
  sf::st_drop_geometry() \%>\%
  dplyr::select(-tidyselect::starts_with("Cost_"), -"cellID", -"metric") \%>\%
  names()

p1 <- prioritizr::problem(out_sf, usedFeatures, "Cost_None") \%>\%
  prioritizr::add_min_set_objective() \%>\%
  prioritizr::add_relative_targets(target$target) \%>\%
  prioritizr::add_binary_decisions() \%>\%
  prioritizr::add_default_solver(verbose = FALSE)

dat_solnClim <- prioritizr::solve.ConservationProblem(p1)
splnr_plot_climKernelDensity_Basic(dat_solnClim)
}
