% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splnr_plotting.R
\name{splnr_plot_featureNo}
\alias{splnr_plot_featureNo}
\title{Plot number of features}
\usage{
splnr_plot_featureNo(df, landmass)
}
\arguments{
\item{df}{An \code{sf} object of features}

\item{landmass}{An \code{sf} object of land polygon}
}
\value{
A ggplot object of the plot
}
\description{
Plot number of features
}
