% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splnr_plotting.R
\name{splnr_plot_impScoreRCPlot}
\alias{splnr_plot_impScoreRCPlot}
\title{Replacement cost importance score (takes too long to use in App)}
\usage{
splnr_plot_impScoreRCPlot(
  soln,
  pDat,
  plotTitle = "",
  colorMap = "A",
  legendTitle = "Importance Score \\n(Replacement Cost Score)"
)
}
\arguments{
\item{soln}{The \code{prioritizr} solution}

\item{pDat}{The \code{prioritizr} problem}

\item{plotTitle}{A character value for the title of the plot. Can be empty ("").}

\item{colorMap}{A character string indicating the color map to use (see https://ggplot2.tidyverse.org/reference/scale_viridis.html for all options)}

\item{legendTitle}{A character value for the title of the legend. Can be empty ("").}
}
\value{
A ggplot object of the plot
}
\description{
Replacement cost importance score (takes too long to use in App)
}
\examples{
\dontrun{
dat_problem <- prioritizr::problem(dat_species_bin \%>\% dplyr::mutate(Cost = runif(n = dim(.)[[1]])),
                                   features = c("Spp1", "Spp2", "Spp3", "Spp4", "Spp5"),
                                   cost_column = "Cost") \%>\%
  prioritizr::add_min_set_objective() \%>\%
  prioritizr::add_relative_targets(0.3) \%>\%
  prioritizr::add_binary_decisions() \%>\%
  prioritizr::add_default_solver(verbose = FALSE)

dat_soln <- dat_problem \%>\%
  prioritizr::solve.ConservationProblem()

(splnr_plot_impScoreRCPlot(dat_soln, dat_problem))
}
}
