% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splnr_plotting.R
\name{splnr_plot_longhurst}
\alias{splnr_plot_longhurst}
\title{Plot Longhurst Provinces}
\usage{
splnr_plot_longhurst(PlanUnits, landmass)
}
\arguments{
\item{PlanUnits}{Planning Units as an \code{sf} object}

\item{landmass}{An \code{sf} object of land polygon}
}
\value{
A ggplot object of the plot
}
\description{
Plot Longhurst Provinces
}
\examples{
\dontrun{
df <- splnr_plot_longhurst(PlanUnits, landmass)
}
}
