% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splnr_plotting.R
\name{splnr_plot_MPAs}
\alias{splnr_plot_MPAs}
\title{Plot MPAs}
\usage{
splnr_plot_MPAs(df, landmass)
}
\arguments{
\item{df}{An \code{sf} object of marine protected areas}

\item{landmass}{An \code{sf} object of land polygon}
}
\value{
A ggplot object of the plot
}
\description{
Plot MPAs
}
