% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splnr_plotting.R
\name{splnr_plot_PUs}
\alias{splnr_plot_PUs}
\title{Plot Planning Units}
\usage{
splnr_plot_PUs(PlanUnits)
}
\arguments{
\item{PlanUnits}{Planning Units as an \code{sf} object}
}
\value{
A ggplot object of the plot
}
\description{
\code{splnr_plot_PUs()} allows to plot the planning units of a planning region (for example created with the \code{spatialplanr}function \code{\link[=splnr_get_planningUnits]{splnr_get_planningUnits()}}) in a customisable way using \code{ggplot2}. This function requires an \code{sf} object containing the geographic information of PUs in the planning region and outputs a \code{ggobject}. It can be combined with the \code{spatialplanr} function \code{\link[=gg_add]{gg_add()}}.
}
\examples{
splnr_plot_PUs(dat_PUs)
}
