% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splnr_plotting.R
\name{splnr_plot_selectionFreq}
\alias{splnr_plot_selectionFreq}
\title{Plot Planning Unit Selection Frequency}
\usage{
splnr_plot_selectionFreq(
  selFreq,
  plotTitle = "",
  paletteName = "Greens",
  legendTitle = "Selection \\nFrequency"
)
}
\arguments{
\item{selFreq}{An \code{sf} object containing the selection frequency data for planning units.
This object must include a \code{selFreq} column (e.g., generated by \code{splnr_get_selFreq()}).}

\item{plotTitle}{A character string for the title of the plot. Defaults to \code{""}.}

\item{paletteName}{A character string or numeric value specifying the name of the
\code{RColorBrewer} palette to use for the fill. Available palettes can be found at
\url{https://ggplot2.tidyverse.org/reference/scale_brewer.html}.
Defaults to \code{"Greens"}.}

\item{legendTitle}{A character string for the title of the legend.
Defaults to \code{"Selection \\nFrequency"}.}
}
\value{
A \code{ggplot} object representing the plot of planning unit selection frequency.
}
\description{
The \code{splnr_plot_selectionFreq()} function visualizes the selection frequency
of planning units across an array of \code{prioritizr} solutions. This is useful
for understanding which areas are consistently selected as important for
conservation.
}
\details{
When multiple spatial plans are generated (either from solutions to different
conservation problems or via a \code{prioritizr} portfolio approach), it's
valuable to assess the robustness of planning unit selection. This function
takes an \code{sf} object as input, which must contain a \code{selFreq} column
representing the selection frequency of each planning unit. This \code{selFreq}
column can be generated using the \code{spatialplanr} function \code{splnr_get_selFreq()}.

The function uses \code{ggplot2} to create a spatial plot of these frequencies,
allowing for customization of the color palette, plot title, and legend title.
The output is a \code{ggplot} object that can be further enhanced by combining it
with the \code{spatialplanr} function \code{splnr_gg_add()}.
}
\examples{
\dontrun{
# Assuming 'dat_species_bin' is an existing sf object in your package.

# Create a dummy prioritizr problem.
dat_problem <- prioritizr::problem(
  dat_species_bin \%>\% dplyr::mutate(Cost = runif(n = dim(.)[[1]])),
  features = c("Spp1", "Spp2", "Spp3", "Spp4", "Spp5"),
  cost_column = "Cost"
) \%>\%
  prioritizr::add_min_set_objective() \%>\%
  prioritizr::add_relative_targets(0.3) \%>\%
  prioritizr::add_binary_decisions() \%>\%
  prioritizr::add_default_solver(verbose = FALSE)

# Create a conservation problem that contains a portfolio of solutions (e.g., 5 solutions).
dat_soln_portfolio <- dat_problem \%>\%
  prioritizr::add_cuts_portfolio(number_solutions = 5) \%>\%
  prioritizr::solve.ConservationProblem()

# Calculate selection frequency using splnr_get_selFreq().
selFreq_data <- splnr_get_selFreq(solnMany = dat_soln_portfolio, type = "portfolio")

# Plot the selection frequency.
plot_selection_frequency <- splnr_plot_selectionFreq(selFreq_data)
print(plot_selection_frequency)
}
}
