% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splnr_plotting.R
\name{splnr_plot_SolutionZones}
\alias{splnr_plot_SolutionZones}
\title{Plot prioritizr solution with zones}
\usage{
splnr_plot_SolutionZones(
  soln,
  colorVals = c("#c6dbef", "#3182bd", "black"),
  legendLabels = c("Not selected", "Zone 1", "Zone 2"),
  showLegend = TRUE,
  plotTitle = "Solution",
  legendTitle = "Planning Units"
)
}
\arguments{
\item{soln}{The \code{prioritizr} solution that has several solutions based on zones}

\item{colorVals}{A \code{list} object of named vectors that will match the color value with the according name. Needs to be number of zones + 1 (to also have a colour for planning units that were not selected).}

\item{legendLabels}{Character values (number of zones + 1) of what the legend should be labelled.}

\item{showLegend}{A logical command on whether to show the legend of the solution (Default: TRUE).}

\item{plotTitle}{A character value for the title of the plot. Can be empty ("").}

\item{legendTitle}{A character value for the title of the legend. Can be empty ("").}
}
\value{
A ggplot object of the plot
}
\description{
Plot prioritizr solution with zones
}
\examples{
t2 <- matrix(NA, ncol = 2, nrow = 5) # create targets
t2[, 1] <- 0.1
t2[, 2] <- 0.05

z2 <- prioritizr::zones("zone 1" = c("Spp1", "Spp2", "Spp3", "Spp4", "Spp5"),
"zone 2" = c("Spp1", "Spp2", "Spp3", "Spp4", "Spp5"))

p2 <- prioritizr::problem(dat_species_bin \%>\% dplyr::mutate(Cost1 = runif(n = dim(.)[[1]]), #when giving sf input, we need as many cost columns as we have zones
                                                            Cost2 = runif(n = dim(.)[[1]])),
                          z2,
                         cost_column = c("Cost1", "Cost2"))\%>\%
  prioritizr::add_min_set_objective() \%>\%
  prioritizr::add_relative_targets(t2) \%>\%
  prioritizr::add_binary_decisions() \%>\%
  prioritizr::add_default_solver(verbose = FALSE)

s2 <- p2 \%>\%
 prioritizr::solve.ConservationProblem()

(splnr_plot_SolutionZones(s2))
}
