% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{splnr_prep_selFreq}
\alias{splnr_prep_selFreq}
\title{Prepare data to plot Selection Frequency of planning units}
\usage{
splnr_prep_selFreq(solnMany, type = "portfolio")
}
\arguments{
\item{solnMany}{List or portfolio of \code{prioritizr} solutions}

\item{type}{Either "portfolio" (\code{sf} object) with a portfolop produced using \code{prioritizr} or "list" with a list of solutions}
}
\value{
\code{selFreq} \code{sf} object containing a column with the selection frequency (sum over all solutions).
}
\description{
Prepare data to plot Selection Frequency of planning units
}
\examples{
\dontrun{
dat_soln_portfolio <- dat_problem \%>\%
  prioritizr::add_top_portfolio(number_solutions = 5) \%>\% #create conservation problem that contains a portfolio of solutions
  prioritizr:::solve.ConservationProblem()

selFreq <- splnr_prep_selFreq(solnMany = dat_soln_portfolio, type = "portfolio")
(splnr_plot_selectionFreq(selFreq))
}
solnList <- list(dat_soln, dat_soln2)
selFreq <- splnr_prep_selFreq(solnMany = solnList, type = "list")
}
