% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{splnr_prepKappaCorrData}
\alias{splnr_prepKappaCorrData}
\title{Prepare data to plot Cohen's Kappa correlation matrix}
\usage{
splnr_prepKappaCorrData(sol, name_sol)
}
\arguments{
\item{sol}{List of \code{prioritizr} solutions (\code{sf} objects) with solutions having a column name \code{solution_1}}

\item{name_sol}{Name tags to the different solutions}
}
\value{
\code{matrixOut} matrix
}
\description{
Prepare data to plot Cohen's Kappa correlation matrix
}
\examples{
corrMat <- splnr_prepKappaCorrData(list(dat_soln, dat_soln2), name_sol = c("soln1", "soln2"))
}
