% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{splnr_prepTargetData}
\alias{splnr_prepTargetData}
\title{Prepare data to plot how well targets are met}
\usage{
splnr_prepTargetData(
  soln,
  pDat,
  allDat,
  Category = NA,
  Dict = NA,
  dataCol = "Abbrev",
  climsmart = FALSE,
  solnCol = "solution_1"
)
}
\arguments{
\item{soln}{The \code{prioritizr} solution}

\item{pDat}{The \code{prioritizr} problem}

\item{allDat}{\code{sf} dataframe containing all features}

\item{Category}{\code{tibble} with information what category the features belong to (if NA, either provide a Dict or each feature has its own category)}

\item{Dict}{\code{tibble} resembling a dictionary that has to contain: Common species name ('Common'), the category that features belong to ('Category'), whether the feature is important or representative ('Class') and what the features are called in the Data (e.g. 'Scientific')  (if NA, either provide Catgories or each feature has its own category)}

\item{dataCol}{Name of the column in the Dictionary (Dict) of what Species names are saved in (default: 'Abbrev')}

\item{climsmart}{logical denoting whether spatial planning was done climate-smart (and targets have to be calculated differently)}

\item{solnCol}{Name of the column with the solution}
}
\value{
\code{tbl_df} dataframe
}
\description{
Prepare data to plot how well targets are met
}
\examples{
\dontrun{
#not including incidental species coverage
dfNInc <- splnr_prepTargetData(soln = dat_soln,
                               pDat = dat_problem,
                               allDat = dat_species_bin,
                               Dict = NA,
                               Category = Category_vec,
                               solnCol = "solution_1")

#including incidental species coverage
dfInc <- splnr_prepTargetData(soln = dat_soln,
                              pDat = dat_problem,
                              allDat = dat_species_bin2,
                              Dict = NA,
                              Category = Category_vec2,
                              solnCol = "solution_1")
}
}
