% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{SpatPlan_Create_Polygon}
\alias{SpatPlan_Create_Polygon}
\title{Function for creating polygon}
\usage{
SpatPlan_Create_Polygon(x, cCRS)
}
\arguments{
\item{x}{}
}
\value{

}
\description{
Function for creating polygon
}
