% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatPlan_Get_AquaMaps.R
\name{SpatPlan_Get_AquaMaps}
\alias{SpatPlan_Get_AquaMaps}
\title{Code to reduce AquaMaps to our requested species and area.}
\usage{
SpatPlan_Get_AquaMaps(
  PlanUnits,
  MinD = 0,
  MaxD = 200,
  Direc = file.path("~", "SpatPlan_Data")
)
}
\arguments{
\item{PlanUnits}{Planning Units as an \code{sf} object}

\item{MinD}{The minimum depth for AquaMaps data}

\item{MaxD}{The maximum depth for AquaMaps data}

\item{Direc}{The directory where the MME data is being stored. If not specified, the default location is assumed.}
}
\value{
An \code{sf} object of AquaMaps data
}
\description{
Code to reduce AquaMaps to our requested species and area.
}
