% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatPlan_Get_AquaMaps.R
\name{SpatPlan_Get_AquaMaps}
\alias{SpatPlan_Get_AquaMaps}
\title{Code to reduce AquaMaps to our requested species and area.}
\usage{
SpatPlan_Get_AquaMaps(
  PlanUnits,
  MinD = 0,
  MaxD = 200,
  CutOff = 0.5,
  limits = NA,
  Direc = file.path("~", "SpatPlan_Data")
)
}
\arguments{
\item{limits}{}
}
\value{

}
\description{
Code to reduce AquaMaps to our requested species and area.
}
