% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatPlan_Get_Boundary.R
\name{SpatPlan_Get_Boundary}
\alias{SpatPlan_Get_Boundary}
\title{Get the boundary of the planning region.}
\usage{
SpatPlan_Get_Boundary(Limits, cCRS, Direc = file.path("~", "SpatPlan_Data"))
}
\arguments{
\item{cCRS}{}
}
\value{

}
\description{
Get the boundary of the planning region.
}
