% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatPlan_Get_Boundary.R
\name{SpatPlan_Get_Boundary}
\alias{SpatPlan_Get_Boundary}
\title{Get the boundary of the planning region.}
\usage{
SpatPlan_Get_Boundary(
  Limits,
  Type,
  cCRS = "+proj=moll +lon_0=0 +x_0=0 +y_0=0 +ellps=WGS84 +datum=WGS84 +units=m no_defs"
)
}
\arguments{
\item{Limits}{The limits of the boundary. This can either be a 4 element numeric named vector (c(xmin = 150, xmax = 160, ymin = -40, ymax = -30)), a vector of ocean/sea names, or a vector of EEZs.,}

\item{Type}{The type of Limits being provided. Options are "Ocean" or "EEZ"}

\item{cCRS}{The CRS the boundary is to be returned in}
}
\value{
The boundary of the planning region
}
\description{
Get the boundary of the planning region.
}
\examples{
Bndry <- SpatPlan_Get_Boundary("Australia", "EEZ")

}
