% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatPlan_Get_FishCost.R
\name{SpatPlan_Get_FishCost}
\alias{SpatPlan_Get_FishCost}
\title{Get fisheries cost layer.}
\usage{
SpatPlan_Get_FishCost(
  PUs,
  group = "all",
  Direc = file.path("~", "SpatPlan_Data")
)
}
\arguments{
\item{Direc}{}
}
\value{

}
\description{
Get fisheries cost layer.
}
\examples{
#' @importFrom rlang .data
}
