% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatPlan_Get_MPAs.R
\name{SpatPlan_Get_MPAs}
\alias{SpatPlan_Get_MPAs}
\title{Get locked in areas}
\usage{
SpatPlan_Get_MPAs(
  PUs,
  Status = "Designated",
  Desig = c("National", "Regional", "International", "Not Applicable"),
  Direc = file.path("~", "SpatPlan_Data")
)
}
\arguments{
\item{PUs}{The planning units}

\item{Status}{The status of the MPA. Options are: "Designated", "Adopted", "Inscribed", "Established"}

\item{Desig}{The designation of the MPA. Options are "National", "Regional", "International", "Not Applicable"}

\item{Direc}{The directory where the MME data is being stored. If not specified, the default location is assumed.}
}
\value{

}
\description{
Get locked in areas
}
\examples{

}
