% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatPlan_Match_Longhurst.R
\name{SpatPlan_Match_Longhurst}
\alias{SpatPlan_Match_Longhurst}
\title{Assign Longhurst Provinces to PUs}
\usage{
SpatPlan_Match_Longhurst(PUs, Direc = file.path("~", "SpatPlan_Data"))
}
\arguments{
\item{Direc}{}
}
\value{

}
\description{
Assign Longhurst Provinces to PUs
}
