% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatPlan_Match_Longhurst.R
\name{SpatPlan_Match_Longhurst}
\alias{SpatPlan_Match_Longhurst}
\title{Assign Longhurst Provinces to PUs}
\usage{
SpatPlan_Match_Longhurst(PlanUnits, Direc = file.path("~", "SpatPlan_Data"))
}
\arguments{
\item{PlanUnits}{Planning Units as an \code{sf} object}

\item{Direc}{The directory where the MME data is being stored. If not specified, the default location is assumed.}
}
\value{
An \code{sf} dataframe with additional longhurst columns (\code{ProvCode}, \code{ProvDescr})
}
\description{
A function that assigns the Longhurst province and description to each planning unit.
}
