% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatPlan_Plotting.R
\name{SpatPlan_Plot_Comparison}
\alias{SpatPlan_Plot_Comparison}
\title{Plot solution comparison}
\usage{
SpatPlan_Plot_Comparison(soln1, soln2, landmass)
}
\arguments{
\item{soln1}{The first \code{prioritizr} solution}

\item{soln2}{The second \code{prioritizr} solution}

\item{landmass}{An \code{sf} object of land polygon}
}
\value{
A ggplot object of the plot
}
\description{
Plot solution comparison
}
