% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatPlan_Plotting.R
\name{SpatPlan_Plot_FeatureNo}
\alias{SpatPlan_Plot_FeatureNo}
\title{Plot number of features}
\usage{
SpatPlan_Plot_FeatureNo(df, landmass)
}
\arguments{
\item{df}{An \code{sf} object of features}

\item{landmass}{An \code{sf} object of land polygon}
}
\value{
A ggplot object of the plot
}
\description{
Plot number of features
}
