% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{spDataFiltered}
\alias{spDataFiltered}
\title{Aquamaps species example data}
\format{
\subsection{\code{spDataFiltered}}{

A data frame with 397 rows and 17 columns:
\describe{
\item{Chelonia_mydas}{suitable habitat for Chelonia mydas}
\item{cellID}{cell ID of planning units}
...
}
}
}
\source{
Aquamaps.org
}
\usage{
spDataFiltered
}
\description{
sf file containing the suitable habitat for a few marine species of the coral sea
}
\keyword{datasets}
