% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{spDataFiltered}
\alias{spDataFiltered}
\title{Example data}
\format{
\code{spDataFiltered}
A data frame with sample species data
\describe{
\item{Species}{Species names}
}
}
\source{
"Made up data"
}
\usage{
spDataFiltered
}
\description{
Example data
}
\keyword{datasets}
