% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splnr_apply_cutoffs.R
\name{splnr_apply_cutoffs}
\alias{splnr_apply_cutoffs}
\title{Function to apply cutoffs to feature data}
\usage{
splnr_apply_cutoffs(features, Cutoffs, inverse = FALSE)
}
\arguments{
\item{features}{A sf dataframe with all the feature information}

\item{Cutoffs}{A single value or a named vector of cutoffs.}

\item{inverse}{If TRUE, values below the \code{Cutoffs} are used. If FALSE (default), values above are kept.}
}
\value{
A new sf dataframe that has cutoffs applied.
}
\description{
Function to apply cutoffs to feature data
}
