% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{splnr_arrangeFeatures}
\alias{splnr_arrangeFeatures}
\title{Ensure all features are in the same order.}
\usage{
splnr_arrangeFeatures(df)
}
\arguments{
\item{df}{An sf object to sort by Lon and Lat}
}
\value{
A sorted sf object with the additionl cellID column
}
\description{
\code{splnr_arrangeFeatures()} sorts your data based on longitude and latitude values.
}
\examples{
df <- dat_species_prob \%>\%
  splnr_arrangeFeatures()
}
