% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-climate.R
\name{splnr_climate_featureApproach}
\alias{splnr_climate_featureApproach}
\title{Function to run the feature approach}
\usage{
splnr_climate_featureApproach(
  features,
  metric,
  targets,
  direction,
  percentile = 35,
  refugiaTarget = 0.3
)
}
\arguments{
\item{features}{feature \code{sf}object}

\item{metric}{climate metric \code{sf} object with 'metric' as the column name of the metric values per planning unit.}

\item{targets}{\code{data.frame}with list of features under "feature" column and their corresponding targets under "target" column}

\item{direction}{If direction = 1, metric values are from low (least climate-smart) to high (most climate-smart). If direction = -1, metric values are from high (least climate-smart) to low (most climate-smart).}

\item{percentile}{cut-off threshold for determining whether an area is a climate priority area or not (e.g., lower 35th percentile of warming or upper 65th percentile of acidification). Note that the percentile here is the lower limit of the threshold.}

\item{refugiaTarget}{target assigned to climate-smart areas}
}
\value{
A \code{list} with two components: 1. is the data frame passed to \code{prioritizr} when creating a conservation problem containing the binary information per planning unit per feature. 2. are the targets for the features in the conservation problem when the CPA approach is used.
}
\description{
Function to run the feature approach
}
\examples{
Features <- dat_species_bin

targets <- Features \%>\%
  sf::st_drop_geometry() \%>\%
  colnames() \%>\%
  data.frame() \%>\%
  setNames(c("feature")) \%>\%
  dplyr::mutate(target = 0.3)

Feature_Approach <- splnr_climate_featureApproach(
  features = dat_species_bin,
  metric = dat_clim,
  targets = targets,
  direction = 1
)
out_sf <- Feature_Approach$Features
targets <- Feature_Approach$Targets
}
