% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{splnr_create_polygon}
\alias{splnr_create_polygon}
\title{Function for creating polygon}
\usage{
splnr_create_polygon(x, cCRS = "EPSG:4326")
}
\arguments{
\item{x}{A named vector of lon/lat coordinates from which to make an \code{sf} polygon}

\item{cCRS}{The CRS to use for the polygon}
}
\value{
An \code{sf} object for the polygon
}
\description{
Function for creating polygon
}
\examples{
splnr_create_polygon(x = dplyr::tibble(x = seq(-50, 50, by = 1), y = 120) \%>\%
                      dplyr::bind_rows(dplyr::tibble(x = 50, y = seq(120, 180, by = 1))) \%>\%
                      dplyr::bind_rows(dplyr::tibble(x = seq(50, -50, by = -1), y = 180)) \%>\%
                      dplyr::bind_rows(dplyr::tibble(x = -50, y = seq(150, 120, by = -1))))
}
