% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splnr_get_boundary.R
\name{splnr_get_boundary}
\alias{splnr_get_boundary}
\title{Get the boundary of the planning region.}
\usage{
splnr_get_boundary(Limits, Type, res = 1, cCRS = "ESRI:54009")
}
\arguments{
\item{Limits}{The limits of the boundary. This can either be a 4 element numeric named vector (c(xmin = 150, xmax = 160, ymin = -40, ymax = -30)), a vector of ocean/sea names, or a vector of EEZs.,}

\item{Type}{The type of Limits being provided. Options are "Ocean" or "EEZ"}

\item{res}{The resolution (in degrees) from which to create the boundary polygon if numeric limits are provided.}

\item{cCRS}{The CRS the boundary is to be returned in}
}
\value{
The boundary of the planning region
}
\description{
Get the boundary of the planning region.
}
\examples{
Bndry <- splnr_get_boundary("North Atlantic Ocean", "Ocean")
}
