% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splnr_get_distCoast.R
\name{splnr_get_distCoast}
\alias{splnr_get_distCoast}
\title{Function to compute distances to nearest coastline for each centroid of each planning unit in the 'sf' object provided.}
\usage{
splnr_get_distCoast(dat_sf)
}
\arguments{
\item{dat_sf}{An sf object.}
}
\value{
An \code{sf} object with distances to the nearest coast
}
\description{
The code takes a sf object and return it updated with a new coastDistance column.
The output inherits the crs from this sf object so ensure it is in the correct projection for your needs
}
\details{
Written by Kristine Buenafe
Written: March/April 2023
Modified by Kilian Barreiro
Updated: December 2023
}
\examples{
# Example 1
crs <- "ESRI:54009" #Mollweide projection
grid <- st_sf(geometry = st_sfc(st_polygon(list(rbind(c(0, 0), c(1, 0), c(1, 1), c(0, 1), c(0, 0)))))) \%>\% 
st_set_crs(crs) \%>\%
st_cast("POLYGON") \%>\%
mutate(cellID = row_number())
splnr_get_distCoast(grid)

# Example 2
crs_2 <- st_crs("+proj=longlat +datum=WGS84 +no_defs")
bbox <- st_bbox(c(xmin = 0, ymin = 0, xmax = 3, ymax = 3))
grid_2 <- st_make_grid(bbox, n = c(3, 3), what = "polygons") 
grid_2 <- st_sf(cellID = 1:length(grid_2), geometry = grid_2) \%>\%
st_set_crs(crs_2)
splnr_get_distCoast(grid_2)
}
