% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splnr_get_MPAs.R
\name{splnr_get_MPAs}
\alias{splnr_get_MPAs}
\title{Get marine parks from the WDPA.}
\usage{
splnr_get_MPAs(
  PlanUnits,
  Countries,
  Status = c("Designated", "Established", "Inscribed"),
  Desig = c("National", "Regional", "International", "Not Applicable"),
  Category = c("Ia", "Ib", "II", "III", "IV"),
  ...
)
}
\arguments{
\item{PlanUnits}{Planning Units as an \code{sf} object}

\item{Countries}{A character vector of the countries for which to extract MPAs. To get all MPAs, use \code{"global"} here.}

\item{Status}{The status field in the WDPA provides information on whether a protected area has been established, designated, or proposed at the time the data was submitted.}

\item{Desig}{The designation type is the category or type of protected area as legally/officially designated or proposed.}

\item{Category}{Stores the IUCN Protected Area Management Categories (recorded in field IUCN_CAT) for each of the protected areas where these categories are reported}

\item{...}{Other arguments passed to \code{wdpa_fetch()}}
}
\value{
A \code{sf} object with the MPAs intersected with the planning units
}
\description{
This code is a wrapper for the wonderful \code{wdpar} package written by Jeffrey O. Hanson. This data is then interfaced with the planning units.
An \code{sf} object is returned with the PU area covered by the selected marine protected areas.
}
\examples{
dat <- splnr_get_MPAs(dat_PUs, "Australia")

aust <- rnaturalearth::ne_countries(country = "Australia", returnclass = "sf")

gg <- ggplot2::ggplot() +
  ggplot2::geom_sf(data = dat, ggplot2::aes(fill = wdpa)) +
  ggplot2::geom_sf(data = aust, fill = "grey50")
}
