% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{splnr_get_selFreq}
\alias{splnr_get_selFreq}
\title{Prepare data to plot Selection Frequency of planning units}
\usage{
splnr_get_selFreq(solnMany, type = "portfolio")
}
\arguments{
\item{solnMany}{List or portfolio of \code{prioritizr} solutions}

\item{type}{Either "portfolio" (\code{sf} object) with a portfolio produced using \code{prioritizr} or "list" with a list of solutions}
}
\value{
\code{selFreq} \code{sf} object containing a column with the selection frequency (sum over all solutions).
}
\description{
Prepare data to plot Selection Frequency of planning units
}
\examples{
dat_problem <- prioritizr::problem(dat_species_bin \%>\% dplyr::mutate(Cost = runif(n = dim(.)[[1]])),
  features = c("Spp1", "Spp2", "Spp3", "Spp4", "Spp5"),
  cost_column = "Cost"
) \%>\%
  prioritizr::add_min_set_objective() \%>\%
  prioritizr::add_relative_targets(0.3) \%>\%
  prioritizr::add_binary_decisions() \%>\%
  prioritizr::add_default_solver(verbose = FALSE)

dat_soln <- dat_problem \%>\%
  prioritizr::solve.ConservationProblem()

# create conservation problem that contains a portfolio of solutions
dat_soln_portfolio <- dat_problem \%>\%
  prioritizr::add_cuts_portfolio(number_solutions = 5) \%>\%
  prioritizr::solve.ConservationProblem()

selFreq <- splnr_get_selFreq(solnMany = dat_soln_portfolio, type = "portfolio")
(splnr_plot_selectionFreq(selFreq))

}
