% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splnr_plotting_climate.R
\name{splnr_plot_climData}
\alias{splnr_plot_climData}
\title{Plot climate data}
\usage{
splnr_plot_climData(
  df,
  colInterest,
  PlanUnits,
  landmass = NA,
  colorMap = "C",
  colorPUs = "grey80",
  plotTitle = " ",
  legendTitle = "Climate metric"
)
}
\arguments{
\item{df}{An \code{sf} object with climate metric information with}

\item{colInterest}{column of data frame that contains the metric informatin}

\item{PlanUnits}{Planning Units as an \code{sf} object}

\item{landmass}{An \code{sf} object of land polygon}

\item{colorMap}{A character string indicating the color map to use (see https://ggplot2.tidyverse.org/reference/scale_viridis.html for all options)}

\item{colorPUs}{A color value for the outline of planning units.}

\item{plotTitle}{A character value for the title of the plot. Can be empty ("").}

\item{legendTitle}{A character value for the title of the legend. Can be empty ("").}
}
\value{
A ggplot object of the plot
}
\description{
Plot climate data
}
\examples{
splnr_plot_climData(dat_clim, dat_clim$metric, dat_PUs)
}
