% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splnr_plotting.R
\name{splnr_plot_CorrMat}
\alias{splnr_plot_CorrMat}
\title{Plot correlation matrices}
\usage{
splnr_plot_CorrMat(
  x,
  colourGradient = c("#BB4444", "#FFFFFF", "#4477AA"),
  legendTitle = "Correlation \\ncoefficient",
  AxisLabels = NULL,
  plotTitle = ""
)
}
\arguments{
\item{x}{A correlation matrix of \code{prioritizr} solutions}

\item{colourGradient}{A list of three colour values for high positive, no and high negative correlation}

\item{legendTitle}{A character value for the title of the legend. Can be empty ("").}

\item{AxisLabels}{A list of labels of the solutions to be correlated (Default: NULL). Length needs to match number of correlated solutions.}

\item{plotTitle}{A character value for the title of the plot. Can be empty ("").}
}
\value{
A ggplot object of the plot
}
\description{
Plot correlation matrices
}
\examples{
CorrMat <- splnr_prepKappaCorrData(list(dat_soln, dat_soln2), name_sol = c("soln1", "soln2"))
(splnr_plot_CorrMat(CorrMat, AxisLabels = c("Solution 1", "Solution 2")))
}
