% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splnr_plotting.R
\name{splnr_plot_cost}
\alias{splnr_plot_cost}
\title{Plot cost}
\usage{
splnr_plot_cost(
  df,
  col_names = NULL,
  legendTitle = "Cost",
  plotTitle = "",
  paletteName = "YlGnBu"
)
}
\arguments{
\item{df}{An \code{sf} object of cost for \code{prioritizr}}

\item{col_names}{Name of the cost column(s)}

\item{legendTitle}{A character value for the title of the legend. Can be empty ("").}

\item{plotTitle}{A character value for the title of the plot. Can be empty ("").}

\item{paletteName}{A string (or number) for the color palette to use. Available palettes can be found at https://ggplot2.tidyverse.org/reference/scale_brewer.html.}
}
\value{
A ggplot object of the plot
}
\description{
\code{splnr_plot_cost()} allows to plot cost within each planning units of a planning region in a customisable way using \code{ggplot2}. This function requires an \code{sf} object with a cost column and outputs a \code{ggobject}. It can be combined with the \code{spatialplanr} function \code{\link[=splnr_gg_add]{splnr_gg_add()}}.
}
\examples{
bathymetry_cont <- oceandatr::get_bathymetry(planning_grid = PUs,
keep = FALSE, classify_bathymetry = FALSE)
splnr_plot_cost(df = bathymetry_cont, col_names = "bathymetry",
legendTitle = "bathymetry", paletteName = , plotTitle = "bathymetry") +
  splnr_gg_add(PUs = PUs, Bndry = Bndry, overlay = landmass,
  cropOverlay = PUs, ggtheme = splnr_theme)

bathymetry_bin <- oceandatr::get_bathymetry(planning_grid = PUs,
 keep = FALSE)
splnr_plot_cost(df = bathymetry_bin, col_names = c("epipelagic",
"hadopelagic","abyssopelagic"), legendTitle = "bathymetry levels",
plotTitle = "bathymetry") +
  splnr_gg_add(PUs = PUs, Bndry = Bndry, overlay = landmass,
  cropOverlay = PUs, ggtheme = splnr_theme)

}
