% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splnr_plotting.R
\name{splnr_plot_cost}
\alias{splnr_plot_cost}
\title{Plot cost}
\usage{
splnr_plot_cost(Cost, Cost_name = "Cost", landmass = NA)
}
\arguments{
\item{Cost}{An \code{sf} object of cost for \code{prioritizr}}

\item{Cost_name}{Name of the cost column}

\item{landmass}{An \code{sf} object of land polygon}
}
\value{
A ggplot object of the plot
}
\description{
Plot cost
}
\examples{
dat_soln \%>\%
  dplyr::mutate(Cost = runif(n = dim(.)[[1]])) \%>\%
  splnr_plot_cost()
}
