% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splnr_plotting.R
\name{splnr_plot_cost}
\alias{splnr_plot_cost}
\title{Plot cost}
\usage{
splnr_plot_cost(
  Cost,
  Cost_name = "Cost",
  paletteName = "YlGnBu",
  plotTitle = "Cost (USD)"
)
}
\arguments{
\item{Cost}{An \code{sf} object of cost for \code{prioritizr}}

\item{Cost_name}{Name of the cost column}

\item{paletteName}{A string (or number) for the color palette to use. Available palettes can be found at https://ggplot2.tidyverse.org/reference/scale_brewer.html.}

\item{plotTitle}{A character value for the title of the plot. Can be empty ("").}
}
\value{
A ggplot object of the plot
}
\description{
Plot cost
}
\examples{
dat_cost <- dat_soln \%>\%
  dplyr::mutate(Cost = runif(n = dim(.)[[1]]))

splnr_plot_cost(dat_cost)
}
