% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splnr_plotting.R
\name{splnr_plot_costOverlay}
\alias{splnr_plot_costOverlay}
\title{Plot cost overlay}
\usage{
splnr_plot_costOverlay(
  soln,
  Cost = NA,
  Cost_name = "Cost",
  legendTitle = "Cost",
  plotTitle = "Solution overlaid with cost"
)
}
\arguments{
\item{soln}{The \code{prioritizr} solution}

\item{Cost}{An \code{sf} object of cost for \code{prioritizr}.In case \code{prioritizr}solution does not contain cost, alternative cost object has to be provided here that was used to generate solution (default: NA).}

\item{Cost_name}{Name of the cost column}

\item{legendTitle}{A character value for the title of the legend. Can be empty ("").}

\item{plotTitle}{A character value for the title of the plot. Can be empty ("").}
}
\value{
A ggplot object of the plot
}
\description{
Plot cost overlay
}
\examples{
dat_problem <- prioritizr::problem(dat_species_bin \%>\% dplyr::mutate(Cost = runif(n = dim(.)[[1]])),
                                   features = c("Spp1", "Spp2", "Spp3", "Spp4", "Spp5"),
                                   cost_column = "Cost") \%>\%
  prioritizr::add_min_set_objective() \%>\%
  prioritizr::add_relative_targets(0.3) \%>\%
  prioritizr::add_binary_decisions() \%>\%
  prioritizr::add_default_solver(verbose = FALSE)

dat_soln <- dat_problem \%>\%
 prioritizr::solve.ConservationProblem()

splnr_plot_costOverlay(soln = dat_soln)
}
