% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splnr_plotting.R
\name{splnr_plot_costOverlay}
\alias{splnr_plot_costOverlay}
\title{Plot cost overlay}
\usage{
splnr_plot_costOverlay(
  soln,
  Cost = NA,
  Cost_name = "Cost",
  landmass = NA,
  legendTitle = "Cost",
  plotTitle = "Solution overlaid with cost"
)
}
\arguments{
\item{soln}{The \code{prioritizr} solution}

\item{Cost}{An \code{sf} object of cost for \code{prioritizr}.In case \code{prioritizr}solution does not contain cost, alternative cost object has to be provided here that was used to generate solution (default: NA).}

\item{Cost_name}{Name of the cost column}

\item{landmass}{An \code{sf} object of land polygon}

\item{legendTitle}{A character value for the title of the legend. Can be empty ("").}

\item{plotTitle}{A character value for the title of the plot. Can be empty ("").}
}
\value{
A ggplot object of the plot
}
\description{
Plot cost overlay
}
\examples{
\dontrun{
(splnr_plot_costOverlay(soln = dat_soln))
}
}
