% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splnr_deprecated.R
\name{splnr_plot_featureNo}
\alias{splnr_plot_featureNo}
\title{Plot number of features}
\usage{
splnr_plot_featureNo(
  df,
  showLegend = TRUE,
  paletteName = "YlGnBu",
  plotTitle = "Number of Features",
  legendTitle = "Features"
)
}
\arguments{
\item{df}{An \code{sf} object of features}

\item{showLegend}{A logical command on whether to show the legend of the solution (Default: TRUE).}

\item{paletteName}{A string (or number) for the color palette to use. Available palettes can be found at https://ggplot2.tidyverse.org/reference/scale_brewer.html.}

\item{plotTitle}{A character value for the title of the plot. Can be empty ("").}

\item{legendTitle}{A character value for the title of the legend. Can be empty ("").}
}
\value{
A ggplot object of the plot
}
\description{
\code{splnr_plot_featureNo()} allows you to use \code{ggplot2} to visually inspect the number of features per planning unit that are used as inputs in the conservation problem. When all features are species, this map can be seen as a visualisation of species richness in the planning region.
}
\details{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

This function requires an \code{sf} object with binary information of all features you want to include in the richness plot (\code{0} for absences and \code{1} for presences, for example created from continuous data with the \code{spatialplanr} function \code{\link[=splnr_apply_cutoffs]{splnr_apply_cutoffs()}}). It outputs a \code{ggobject} and can be combined with the \code{spatialplanr} function \code{\link[=splnr_gg_add]{splnr_gg_add()}}.
}
\examples{
\dontrun{
(splnr_plot_featureNo(dat_species_bin))
}
}
