% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splnr_featureRep.R
\name{splnr_plot_featureRep}
\alias{splnr_plot_featureRep}
\title{Plot how well targets are met}
\usage{
splnr_plot_featureRep(
  df,
  category = NA,
  nr = 1,
  showTarget = NA,
  plotTitle = ""
)
}
\arguments{
\item{df}{A \code{df} containing the target information (resulting from the splnr_get_featureRep() function)}

\item{category}{A named charcter vector of feature and category for grouping the plot output}

\item{nr}{Number of rows of the legend}

\item{showTarget}{\code{logical} Should the targets be shown on the bar plot}

\item{plotTitle}{A character value for the title of the plot. Can be empty ("").}
}
\value{
A ggplot object of the plot
}
\description{
Plot how well targets are met
}
\examples{
dat_problem <- prioritizr::problem(dat_species_bin \%>\% dplyr::mutate(Cost = runif(n = dim(.)[[1]])),
                                   features = c("Spp1", "Spp2", "Spp3", "Spp4", "Spp5"),
                                   cost_column = "Cost") \%>\%
  prioritizr::add_min_set_objective() \%>\%
  prioritizr::add_relative_targets(0.3) \%>\%
  prioritizr::add_binary_decisions() \%>\%
  prioritizr::add_default_solver(verbose = FALSE)

dat_soln <- dat_problem \%>\%
 prioritizr::solve.ConservationProblem()


# including incidental species coverage
df <- splnr_get_featureRep(
  soln = dat_soln,
  pDat = dat_problem)

(splnr_plot_featureRep(df, category = category_df))

}
