% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splnr_plotting.R
\name{splnr_plot_impScoreRWRPlot}
\alias{splnr_plot_impScoreRWRPlot}
\title{Plot Rarity weighted richness importance score (can be used with minimum shortfall OBJECTIVE FUNCTION)}
\usage{
splnr_plot_impScoreRWRPlot(
  soln,
  pDat,
  plotTitle = "",
  colorMap = "A",
  legendTitle = "Importance Score \\n(Rarity Weighted Richness Score)"
)
}
\arguments{
\item{soln}{The \code{prioritizr} solution}

\item{pDat}{The \code{prioritizr} problem}

\item{plotTitle}{A character value for the title of the plot. Can be empty ("").}

\item{colorMap}{A character string indicating the color map to use (see https://ggplot2.tidyverse.org/reference/scale_viridis.html for all options)}

\item{legendTitle}{A character value for the title of the legend. Can be empty ("").}
}
\value{
A ggplot object of the plot
}
\description{
Plot Rarity weighted richness importance score (can be used with minimum shortfall OBJECTIVE FUNCTION)
}
\examples{
\dontrun{
(splnr_plot_impScoreRWRPlot(dat_soln, dat_problem))
}
}
