% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splnr_deprecated.R
\name{splnr_plot_MPAs}
\alias{splnr_plot_MPAs}
\title{Plot MPAs}
\usage{
splnr_plot_MPAs(
  df,
  colorVals = c(`TRUE` = "blue", `FALSE` = "white"),
  showLegend = TRUE,
  plotTitle = "Locked In Areas",
  legendTitle = ""
)
}
\arguments{
\item{df}{An \code{sf} object of marine protected areas}

\item{colorVals}{A \code{list} object of named vectors that will match the color value with the according name. "TRUE" stands for selected planning units.}

\item{showLegend}{A logical command on whether to show the legend of the solution (Default: TRUE).}

\item{plotTitle}{A character value for the title of the plot. Can be empty ("").}

\item{legendTitle}{A character value for the title of the legend. Can be empty ("").}
}
\value{
A ggplot object of the plot
}
\description{
\code{splnr_plot_MPAs()} allows to plot either the outline or the area of MPAs existing in the planning region (for example extracted with the \code{spatialplanr}function \code{\link[=splnr_get_MPAs]{splnr_get_MPAs()}}) in a customisable way using \code{ggplot2}. This function requires an \code{sf} object containing the information whether a planning unit in the planning region lies within an MPA or not in a column called \code{wdpa} and outputs a \code{ggobject}. It can be combined with the \code{spatialplanr} function \code{\link[=splnr_gg_add]{splnr_gg_add()}}.
}
\details{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
}
\examples{
\dontrun{
splnr_plot_MPAs(dat_mpas)
}
}
