% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splnr_plotting.R
\name{splnr_plot_PUs}
\alias{splnr_plot_PUs}
\title{Plot Planning Units}
\usage{
splnr_plot_PUs(PlanUnits)
}
\arguments{
\item{PlanUnits}{Planning Units as an \code{sf} object}
}
\value{
A ggplot object of the plot
}
\description{
Plot Planning Units
}
\examples{
splnr_plot_PUs(dat_PUs)
}
