% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splnr_plotting.R
\name{splnr_plot_selectionFreq}
\alias{splnr_plot_selectionFreq}
\title{Plot selection frequency of a planning unit in an array of prioritisations}
\usage{
splnr_plot_selectionFreq(
  selFreq,
  landmass = NA,
  plotTitle = "",
  paletteName = "Greens",
  legendTitle = "Selection \\nFrequency"
)
}
\arguments{
\item{selFreq}{An \code{sf} object containing the selection frequency of a planning unit from an array of solutions}

\item{landmass}{An \code{sf} object of land polygon}

\item{plotTitle}{A character value for the title of the plot. Can be empty ("").}

\item{paletteName}{A string (or number) for the color palette to use. Available palettes can be found at https://ggplot2.tidyverse.org/reference/scale_brewer.html.}

\item{legendTitle}{A character value for the title of the legend. Can be empty ("").}
}
\value{
A ggplot object of the plot
}
\description{
Plot selection frequency of a planning unit in an array of prioritisations
}
\examples{
\dontrun{
dat_soln_portfolio <- dat_problem \%>\%
  prioritizr::add_top_portfolio(number_solutions = 5) \%>\% #create conservation problem that contains a portfolio of solutions
  prioritizr:::solve.ConservationProblem()

selFreq <- dat_soln_portfolio \%>\%  # calculate selection frequency
sf::st_drop_geometry() \%>\%
dplyr::mutate(selFreq = as.factor(rowSums(dplyr::select(., dplyr::starts_with("solution_"))))) \%>\%
sf::st_as_sf(geometry = dat_soln_portfolio$geometry) \%>\%
dplyr::select(selFreq)

(splnr_plot_selectionFreq(selFreq))
}
}
