% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splnr_plotting.R
\name{splnr_plot_Solution}
\alias{splnr_plot_Solution}
\title{Plot prioritizr solution}
\usage{
splnr_plot_Solution(soln, PlanUnits, landmass)
}
\arguments{
\item{soln}{The \code{prioritizr} solution}

\item{PlanUnits}{Planning Units as an \code{sf} object}

\item{landmass}{An \code{sf} object of land polygon}
}
\value{
A ggplot object of the plot
}
\description{
Plot prioritizr solution
}
