% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{splnr_prepKappaCorrData}
\alias{splnr_prepKappaCorrData}
\title{Prepare data to plot Cohen's Kappa correlation matrix}
\usage{
splnr_prepKappaCorrData(sol, name_sol)
}
\arguments{
\item{sol}{List of \code{prioritizr} solutions (\code{sf} objects) with solutions having a column name \code{solution_1}}

\item{name_sol}{Name tags to the different solutions}
}
\value{
\code{matrixOut} matrix
}
\description{
Prepare data to plot Cohen's Kappa correlation matrix
}
\examples{
# 30 \% target for problem/solution 1
dat_problem <- prioritizr::problem(dat_species_bin \%>\% dplyr::mutate(Cost = runif(n = dim(.)[[1]])),
                                   features = c("Spp1", "Spp2", "Spp3", "Spp4", "Spp5"),
                                   cost_column = "Cost") \%>\%
  prioritizr::add_min_set_objective() \%>\%
  prioritizr::add_relative_targets(0.3) \%>\%
  prioritizr::add_binary_decisions() \%>\%
  prioritizr::add_default_solver(verbose = FALSE)

dat_soln <- dat_problem \%>\%
  prioritizr::solve.ConservationProblem()

# 50 \% target for problem/solution 2
dat_problem2 <- prioritizr::problem(dat_species_bin \%>\%
                                      dplyr::mutate(Cost = runif(n = dim(.)[[1]])),
                                   features = c("Spp1", "Spp2", "Spp3", "Spp4", "Spp5"),
                                   cost_column = "Cost") \%>\%
  prioritizr::add_min_set_objective() \%>\%
  prioritizr::add_relative_targets(0.5) \%>\%
  prioritizr::add_binary_decisions() \%>\%
  prioritizr::add_default_solver(verbose = FALSE)

dat_soln2 <- dat_problem2 \%>\%
  prioritizr::solve.ConservationProblem()

corrMat <- splnr_prepKappaCorrData(list(dat_soln, dat_soln2), name_sol = c("soln1", "soln2"))
}
