% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{splnr_replace_NAs}
\alias{splnr_replace_NAs}
\title{Remove NAs from spatial data using nearest neighbour}
\usage{
splnr_replace_NAs(df, vari)
}
\arguments{
\item{df}{An \code{sf} dataframe}

\item{vari}{Variable to remove NAs from}
}
\value{
An \code{sf} object with NAs replaced with the nearest neighbour
}
\description{
\code{splnr_replace_NAs()} allows you to replace NA values in your data with the value of the nearest neighbour.
The nearest neighbour is determined using \code{st_nearest_feature()} from the \code{sf} package.
}
\examples{
df <- dat_species_prob \%>\%
  splnr_replace_NAs("Spp2")
}
