% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{splnr_replace_NAs}
\alias{splnr_replace_NAs}
\title{Remove NAs from spatial data using nearest neighbour}
\usage{
splnr_replace_NAs(df, vari)
}
\arguments{
\item{df}{An \code{sf} dataframe}

\item{vari}{Variable to remove NAs from}
}
\value{
An \code{sf} object with NAs replaced with the nearest neighbour
}
\description{
Remove NAs from spatial data using nearest neighbour
}
\examples{
df <- dat_species_prob \%>\%
  splnr_replace_NAs("Spp2")
}
