% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{splnr_scale_01}
\alias{splnr_scale_01}
\title{Scale spatial layers to between 0 and 1}
\usage{
splnr_scale_01(dat, col_name)
}
\arguments{
\item{dat}{\code{sf} dataframe}

\item{col_name}{Name of the column to scale}
}
\value{
\code{sf} dataframe
}
\description{
Scale spatial layers to between 0 and 1
}
\examples{
df <- dat_species_prob \%>\%
    dplyr::mutate(Spp1 = Spp1 * 100) \%>\%
    splnr_scale_01(col_name = "Spp1")
}
