% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splnr_targets.R
\name{splnr_targets_byCategory}
\alias{splnr_targets_byCategory}
\title{Assign targets to all features by category}
\usage{
splnr_targets_byCategory(dat, catTarg, catName = "Category")
}
\arguments{
\item{dat}{A sf object with the features and categories}

\item{catTarg}{A named character vector with categories and target}

\item{catName}{An optional argument for the name of the category coolumn in dat}
}
\value{
An sf object with targets added
}
\description{
Assign targets to all features by category
}
\examples{
dat <- splnr_targets_byCategory(dat = category_df,
                               catTarg = c("Group1" = 0.5, "Group2" = 0.2),
                               catName = "category")
}
