% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splnr_targets.R
\name{splnr_targets_byIUCN}
\alias{splnr_targets_byIUCN}
\title{Assign targets bu IUCN Red List categories}
\usage{
splnr_targets_byIUCN(dat, IUCN_target, IUCN_col = "IUCN_Category")
}
\arguments{
\item{dat}{A dataframe or sf object with IUCN categories}

\item{IUCN_target}{Either a numeric or named numeric of targets to apply to IUCN categories}

\item{IUCN_col}{Optional string to indicate the name of the column with the IUCN categories}
}
\value{
dataframe or sf object
}
\description{
\code{splnr_targets_byIUCN()} allows to assign targets for species used in conservation planning based on IUCN categories. Species can be extracted based on IUCN categories with the \code{spatoalplnr}function \code{splnr_get_IUCNRedList()}.
Accessing the IUCN database requires a login token from \code{rl_use_iucn()} that needs to be added to the environment using \code{Sys.setenv(IUCN_REDLIST_KEY = "[Your Token]")}. You can start by running \code{rredlist::rl_use_iucn()}.
}
\examples{
dat <- data.frame(IUCN_Category = c("EW", "EX", NA), target = c(0.3, 0.3, 0.3))
IUCN_target <- c("EX" = 0.8, "EW" = 0.6)
dat <- splnr_targets_byIUCN(dat, IUCN_target)
}
